+++ src/javascript/API.js
@@ -488,6 +488,7 @@ const graviton = {
   },
   getPlugin: function(folder_name){
     for(i=0;i<plugins_list.length;i++){
+      console.log(plugins_list[i].folder)
       if(plugins_list[i].folder == folder_name){
         return plugins_list[i];
       }